// src/Types/AddressTypes.ts
export type AddressDto = {
  addressName: string;
  phoneNumber: string;
  country: string;
  city: string;
  firstName: string;
  lastName: string;
  fullAddress: string;
  userId: string;
};

export type AddressResponseDto = {
  id: string;
  addressName: string;
  phoneNumber: string;
  country: string;
  city: string;
  firstName: string;
  lastName: string;
  fullAddress: string;
  userId: string;
};